Refactor and release notes (#109)
