+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include README.rst
 include LICENSE
+include pytest_pylint
