+++ pytest_pylint/tests/test_pytest_pylint.py
@@ -126,26 +126,12 @@ msg-template=start {msg_id} end
     assert 'start W0611 end' in result.stdout.str()
 
 
-def test_get_rel_path():
-    """
-    Verify our relative path function.
-    """
-    from pytest_pylint import get_rel_path
-    correct_rel_path = 'How/Are/You/blah.py'
-    path = '/Hi/How/Are/You/blah.py'
-    parent_path = '/Hi/'
-    assert get_rel_path(path, parent_path) == correct_rel_path
-
-    parent_path = '/Hi'
-    assert get_rel_path(path, parent_path) == correct_rel_path
-
-
 def test_multiple_jobs(testdir):
     """
     Assert that the jobs argument is passed through to pylint if provided
     """
     testdir.makepyfile("""import sys""")
-    with mock.patch('pytest_pylint.lint.Run') as run_mock:
+    with mock.patch('pytest_pylint.plugin.lint.Run') as run_mock:
         jobs = 0
         testdir.runpytest(
             '--pylint', '--pylint-jobs={0}'.format(jobs)
@@ -159,65 +145,10 @@ def test_no_multiple_jobs(testdir):
     If no jobs argument is specified it should not appear in pylint arguments
     """
     testdir.makepyfile("""import sys""")
-    with mock.patch('pytest_pylint.lint.Run') as run_mock:
+    with mock.patch('pytest_pylint.plugin.lint.Run') as run_mock:
         testdir.runpytest('--pylint')
     assert run_mock.call_count == 1
     assert '-j' not in run_mock.call_args[0][0]
-
-
-def test_include_path():
-    """
-    Files should only be included in the list if none of the directories on
-    it's path, of the filename, match an entry in the ignore list.
-    """
-    from pytest_pylint import include_file
-    ignore_list = [
-        "first", "second", "third", "part", "base.py"
-    ]
-    # Default includes.
-    assert include_file("random", ignore_list) is True
-    assert include_file("random/filename", ignore_list) is True
-    assert include_file("random/other/filename", ignore_list) is True
-    # Basic ignore matches.
-    assert include_file("first/filename", ignore_list) is False
-    assert include_file("random/base.py", ignore_list) is False
-    # Part on paths.
-    assert include_file("part/second/filename.py", ignore_list) is False
-    assert include_file("random/part/filename.py", ignore_list) is False
-    assert include_file("random/second/part.py", ignore_list) is False
-    # Part as substring on paths.
-    assert include_file("part_it/other/filename.py", ignore_list) is True
-    assert include_file("random/part_it/filename.py", ignore_list) is True
-    assert include_file("random/other/part_it.py", ignore_list) is True
-
-
-def test_pylint_ignore_patterns():
-    """Test if the ignore-patterns is working"""
-    from pytest_pylint import include_file
-    ignore_patterns = [
-        "first.*",
-        ".*second",
-        "^third.*fourth$",
-        "part",
-        "base.py"
-    ]
-
-    # Default includes
-    assert include_file("random", [], ignore_patterns) is True
-    assert include_file("random/filename", [], ignore_patterns) is True
-    assert include_file("random/other/filename", [], ignore_patterns) is True
-
-    # Pattern matches
-    assert include_file("first1", [], ignore_patterns) is False
-    assert include_file("first", [], ignore_patterns) is False
-    assert include_file("_second", [], ignore_patterns) is False
-    assert include_file("second_", [], ignore_patterns) is False
-    assert include_file("second_", [], ignore_patterns) is False
-    assert include_file("third fourth", [], ignore_patterns) is False
-    assert include_file("_third fourth_", [], ignore_patterns) is True
-    assert include_file("part", [], ignore_patterns) is False
-    assert include_file("1part2", [], ignore_patterns) is True
-    assert include_file("base.py", [], ignore_patterns) is False
 
 
 def test_skip_checked_files(testdir):
