+++ setup.py
@@ -13,12 +13,12 @@ setup(
     description='pytest plugin to check source code with pylint',
     long_description=open("README.rst").read(),
     license='MIT',
-    version='0.14.1',
+    version='0.15.0',
     author='Carson Gee',
     author_email='x@carsongee.com',
     url='https://github.com/carsongee/pytest-pylint',
-    py_modules=['pytest_pylint'],
-    entry_points={'pytest11': ['pylint = pytest_pylint']},
+    packages=['pytest_pylint'],
+    entry_points={'pytest11': ['pylint = pytest_pylint.plugin']},
     python_requires=">=3.5",
     install_requires=['pytest>=5.0', 'pylint>=2.0.0'],
     setup_requires=['pytest-runner'],
